needs refactoring
